Revert "Added ability to set affix type, value, case sensitive while checking Interface name, changed docs"
