+++ src/Standards/Generic/Tests/NamingConventions/InterfaceSuffixRequiredForInterfaceUnitTest.inc
@@ -1,0 +1,27 @@
+<?php
+
+interface SomeNameInterface
+{
+
+}
+
+interface ISomeName // error
+{
+
+}
+
+interface ISomeNameInterface
+{
+
+}
+
+
+interface ISomeOtherNameinterface
+{
+
+}
+
+interface mInterface
+{
+
+}
